% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.r
\name{add}
\alias{add}
\title{Add samples}
\usage{
add(x, y)
}
\arguments{
\item{x}{one number}

\item{y}{one number}
}
\description{
Add samples
}
\section{Functions}{
\itemize{
\item \code{add()}: x and y.

}}
\examples{
add(1,2)
\donttest{
add(1,2)
}
}
